[DEV-1591] Fix for normal mandators
