+++ src/shared/db/entity.ts
@@ -4,11 +4,11 @@ export class IEntity {
   @PrimaryGeneratedColumn()
   id: number;
 
-  @UpdateDateColumn()
-  updated: Date;
-
   @CreateDateColumn()
   created: Date;
+
+  @UpdateDateColumn()
+  updated: Date;
 
   static copy<T extends IEntity>(entity: T): Partial<T> {
     const copy: Partial<T> = { ...entity };
